+++ X10D.Unity/src/Numerics/Vector2IntExtensions.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Runtime.CompilerServices;
 using UnityEngine;
+using X10D.Drawing;
 
 namespace X10D.Unity.Numerics;
 
@@ -20,6 +21,73 @@ public static class Vector2IntExtensions
     {
         x = vector.x;
         y = vector.y;
+    }
+
+    /// <summary>
+    ///     Determines if the current <see cref="Vector2Int" /> lies on the specified <see cref="LineF" />.
+    /// </summary>
+    /// <param name="point">The point to check.</param>
+    /// <param name="line">The line on which the point may lie.</param>
+    /// <returns>
+    ///     <see langword="true" /> if <paramref name="point" /> lies on the line defined by <paramref name="line" />; otherwise
+    ///     <see langword="false" />.
+    /// </returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool IsOnLine(this Vector2Int point, LineF line)
+    {
+        return point.ToSystemPoint().IsOnLine(line);
+    }
+
+    /// <summary>
+    ///     Determines if the current <see cref="Vector2Int" /> lies on the specified line.
+    /// </summary>
+    /// <param name="point">The point to check.</param>
+    /// <param name="start">The starting point of the line.</param>
+    /// <param name="end">The ending point of the line.</param>
+    /// <returns>
+    ///     <see langword="true" /> if <paramref name="point" /> lies on the line defined by <paramref name="start" /> and
+    ///     <paramref name="end" />; otherwise <see langword="false" />.
+    /// </returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool IsOnLine(this Vector2Int point, PointF start, PointF end)
+    {
+        return point.IsOnLine(new LineF(start, end));
+    }
+
+    /// <summary>
+    ///     Determines if the current <see cref="Vector2Int" /> lies on the specified line.
+    /// </summary>
+    /// <param name="point">The point to check.</param>
+    /// <param name="start">The starting point of the line.</param>
+    /// <param name="end">The ending point of the line.</param>
+    /// <returns>
+    ///     <see langword="true" /> if <paramref name="point" /> lies on the line defined by <paramref name="start" /> and
+    ///     <paramref name="end" />; otherwise <see langword="false" />.
+    /// </returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool IsOnLine(this Vector2Int point, Vector2Int start, Vector2Int end)
+    {
+        return point.ToSystemPoint().IsOnLine(new LineF(start.ToSystemVector(), end.ToSystemVector()));
+    }
+
+    /// <summary>
+    ///     Determines if the current <see cref="Vector2Int" /> lies on the specified line.
+    /// </summary>
+    /// <param name="point">The point to check.</param>
+    /// <param name="start">The starting point of the line.</param>
+    /// <param name="end">The ending point of the line.</param>
+    /// <returns>
+    ///     <see langword="true" /> if <paramref name="point" /> lies on the line defined by <paramref name="start" /> and
+    ///     <paramref name="end" />; otherwise <see langword="false" />.
+    /// </returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool IsOnLine(this Vector2Int point, Vector2 start, Vector2 end)
+    {
+        return point.ToSystemPoint().IsOnLine(new LineF(start.ToSystemVector(), end.ToSystemVector()));
     }
 
     /// <summary>